refactor(ethereum): move code into state update
No point in having a wrapper.
