--- crates/ethereum/src/log/fetch.rs
@@ -1,3 +1,0 @@
-mod forward;
-
-pub use forward::*;
